refactor(nav): disambiguate graph node internals
