+++ packages/libraries/graph-tools/bin/vt-graph/commands/apply.ts
@@ -1,4 +1,4 @@
-import {graphStateApply} from '../../../src/node'
+import {graphStateApply} from '../../../src/node-runtime'
 
 export async function runApplyCommand(args: readonly string[]): Promise<void> {
   await graphStateApply(args)
