+++ packages/libraries/graph-tools/bin/vt-graph/commands/mv.ts
@@ -1,4 +1,4 @@
-import {graphMove} from '../../../src/node'
+import {graphMove} from '../../../src/node-runtime'
 
 export async function runMvCommand(args: string[]): Promise<void> {
   await graphMove(0, undefined, args)
