+++ packages/libraries/graph-tools/bin/vt-graph/commands/rename.ts
@@ -1,4 +1,4 @@
-import {graphRename} from '../../../src/node'
+import {graphRename} from '../../../src/node-runtime'
 
 export async function runRenameCommand(args: string[]): Promise<void> {
   await graphRename(0, undefined, args)
