+++ packages/systems/voicetree-cli/src/commands/graph-node/group.ts
@@ -1,2 +1,2 @@
 // Thin wrapper: implementation has moved to @vt/graph-tools
-export { graphMove } from '@vt/graph-tools/node'
+export { graphGroup } from '@vt/graph-tools/node-runtime'
