+++ packages/systems/voicetree-cli/src/commands/graph-node/move.ts
@@ -1,2 +1,2 @@
 // Thin wrapper: implementation has moved to @vt/graph-tools
-export { graphRename } from '@vt/graph-tools/node'
+export { graphMove } from '@vt/graph-tools/node-runtime'
