+++ packages/systems/voicetree-cli/src/commands/graph-node/rename.ts
@@ -1,2 +1,2 @@
 // Thin wrapper: implementation has moved to @vt/graph-tools
-export { graphGroup } from '@vt/graph-tools/node'
+export { graphRename } from '@vt/graph-tools/node-runtime'
