+++ packages/systems/voicetree-cli/src/commands/graph-node/search.ts
