+++ packages/systems/voicetree-cli/src/commands/graph-node/view.ts
