+++ packages/systems/voicetree-cli/src/voicetree-cli.ts
@@ -15,7 +15,7 @@ import {runManualCommand} from './commands/manual.ts'
 import {findRepoRoot} from './commands/util/findRepoRoot.ts'
 import {runSessionCommand} from './commands/runtime/session.ts'
 import {runProjectCommand} from './commands/runtime/project.ts'
-import {runViewCommand} from './commands/node/view.ts'
+import {runViewCommand} from './commands/graph-node/view.ts'
 import {getErrorMessage} from './commands/graph/core/util.ts'
 import {CliError, error} from './commands/output.ts'
 import {argsShape} from './commands/telemetry/argsShape.ts'
@@ -209,17 +209,17 @@ async function dispatchGraphCommand(
             return
         }
         case 'rename': {
-            const {graphRename} = await import('./commands/node/rename.ts')
+            const {graphRename} = await import('./commands/graph-node/rename.ts')
             await graphRename(terminalId, args)
             return
         }
         case 'mv': {
-            const {graphMove} = await import('./commands/node/move.ts')
+            const {graphMove} = await import('./commands/graph-node/move.ts')
             await graphMove(terminalId, args)
             return
         }
         case 'group': {
-            const {graphGroup} = await import('./commands/node/group.ts')
+            const {graphGroup} = await import('./commands/graph-node/group.ts')
             await graphGroup(terminalId, args)
             return
         }
@@ -237,7 +237,7 @@ async function dispatchSearchCommand(
     terminalId: string | undefined,
     args: string[]
 ): Promise<void> {
-    const {searchCommand} = await import('./commands/node/search.ts')
+    const {searchCommand} = await import('./commands/graph-node/search.ts')
     await searchCommand(terminalId, args)
 }
