+++ packages/systems/vt-daemon/src/create-graph/createGraphValidation.ts
@@ -11,7 +11,7 @@ import type {CreateGraphNodeInput} from './createGraphTypes'
 import type {Graph, NodeIdAndFilePath} from '@vt/graph-model/graph'
 import {isFolderIdentityNote} from '@vt/graph-model/graph'
 import {extractParentRefs, normalizeBatchFilenameKey, findBestMatchingNode, type ParentLineRef} from '@vt/graph-model/markdown'
-import {computeGraphComplexity, type EdgePair, type GraphComplexityResult} from '@vt/graph-tools/node'
+import {computeGraphComplexity, type EdgePair, type GraphComplexityResult} from '@vt/graph-tools/node-runtime'
 import {countBodyLines} from '../tools/graph/addProgressNodeTool'
 import {countFolderBoundedComponent, collectFolderBoundedComponent} from './subgraphComponent'
 import * as daemonProtocol from '@vt/vt-daemon-protocol'
